refactor: line cooperation
