+++ cooperation/src/main/scala/com/github/BambooTuna/AkkaServerSupport/cooperation/line/LineAccessTokenAcquisitionResponse.scala
@@ -1,21 +1,12 @@
 package com.github.BambooTuna.AkkaServerSupport.cooperation.line
 
-import com.github.BambooTuna.AkkaServerSupport.cooperation.model.{
-  AccessTokenAcquisitionResponseFailed,
-  AccessTokenAcquisitionResponseSuccess
-}
+import com.github.BambooTuna.AkkaServerSupport.authentication.oauth2.AccessTokenAcquisitionResponse
 
-case class LineAccessTokenAcquisitionResponseSuccess(
+case class LineAccessTokenAcquisitionResponse(
     access_token: String,
     expires_in: Long,
     id_token: Option[String],
     refresh_token: String,
     scope: String,
     token_type: String,
-) extends AccessTokenAcquisitionResponseSuccess
-
-case class LineAccessTokenAcquisitionResponseFailed(
-    error: String,
-    error_description: Option[String],
-    error_uri: Option[String])
-    extends AccessTokenAcquisitionResponseFailed
+) extends AccessTokenAcquisitionResponse
