--- cooperation/src/main/scala/com/github/BambooTuna/AkkaServerSupport/cooperation/model/ClientAuthenticationRequest.scala
@@ -1,9 +1,0 @@
-package com.github.BambooTuna.AkkaServerSupport.cooperation.model
-
-trait ClientAuthenticationRequest {
-  val client_id: String
-  val scope: Option[String]
-  val state: String
-  val redirect_uri: String
-  val response_type: String
-}
