--- cooperation/src/main/scala/com/github/BambooTuna/AkkaServerSupport/cooperation/model/OAuth2Settings.scala
@@ -1,7 +1,0 @@
-package com.github.BambooTuna.AkkaServerSupport.cooperation.model
-
-import com.github.BambooTuna.AkkaServerSupport.core.session.StorageStrategy
-import com.typesafe.config.Config
-
-case class OAuth2Settings(clientConfig: ClientConfig,
-                          strategy: StorageStrategy[String, String])
